feat(api): search database maps are now multivalue
If multiple tweaks share the same translated display name, then only one of the tweaks would appear in the results.

Forcing search databases to be multivalue fixes this issue and allows a unique database key to be associated with multiple objects.

Some parts of the api that utilize databases were updated to be compatible with the new search results logic.

Refs: none